Adding and deleting friends api
To add example:
https://localhost/friends.manip?idVk=13579&idVkFriend=12357&op=add
To delete example:
https://localhost/friends.manip?idVk=13579&idVkFriend=12357&op=del

+++ Server/src/server/logic/FriendDAO.java
@@ -14,7 +14,7 @@ import server.entity.Friend;
  *
  * @author Анастасия Тарасова
  * 
- * changed by oglandx on 04.13.2015
+ * changed by oglandx on 04.13.2015, 21.04.2015
  */
 public class FriendDAO {
     
@@ -22,7 +22,7 @@ public class FriendDAO {
         Session session = HibernateUtil.getSessionFactory().openSession();
         try {
             return (List<Friend>) session.createCriteria(Friend.class)
-                    .add(Restrictions.gt("singleFriend", 0))
+                    .add(Restrictions.gt("singleFriend", 1))
                     .add(Restrictions.eq("idVk", idVk))
                     .list();
         } finally {
@@ -30,35 +30,52 @@ public class FriendDAO {
         }
     }
          
-    public static void addSingleFriend(long idUser, long idFriend) {
-        Friend friend = new Friend();
-        friend.setIdUser(idUser);
-        friend.setSingleFriend(true);
-        friend.setIdFriend(idFriend);
-        
+    public static void addSingleFriend(long idVkUser, long idVkFriend) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         try {
+            Friend friend = (Friend)session
+                    .createCriteria(Friend.class)
+                    .add(Restrictions.eq("singleFriend", 1))
+                    .add(Restrictions.eq("idVk", idVkUser))
+                    .add(Restrictions.eq("idVkFriend", idVkFriend))
+                    .uniqueResult();
+            if(friend != null){
+                return; // friend already created
+            }
+            else{
+                friend = new Friend();
+                friend.setIdVkUser(idVkUser);
+                friend.setSingleFriend(true);
+                friend.setIdVkFriend(idVkFriend);
+            }
             session.beginTransaction();           
             session.save(friend);         
             session.getTransaction().commit();
         } finally {
             session.close();
-        };
+        }
     }
         
             
-    public static void deleteSingleFriend(long idUser, long idFriend){
-        Friend friend = new Friend();
-        friend.setIdUser(idUser);
-        friend.setIdFriend(idFriend);
-        
+    public static void deleteSingleFriend(long idVkUser, long idVkFriend){
         Session session = HibernateUtil.getSessionFactory().openSession();
         try {
-            session.beginTransaction();           
-            session.delete(friend);         
+            List<Friend> friends = (List<Friend>)session
+                    .createCriteria(Friend.class)
+                    .add(Restrictions.eq("singleFriend", 1))
+                    .add(Restrictions.eq("idVk", idVkUser))
+                    .add(Restrictions.eq("idVkFriend", idVkFriend))
+                    .list();
+            if(friends.isEmpty()){
+                return;
+            }
+            session.beginTransaction();
+            for (Friend friend: friends){
+                session.delete(friend);
+            }
             session.getTransaction().commit();
         } finally {
             session.close();
-        };
+        }
     }
 }
